requirements.txt
numpy>=1.24
pandas>=2.0
yfinance>=0.2.40
scikit-learn>=1.4
python-dotenv>=1.0
requests>=2.31
feedparser>=6.0
google-genai>=1.0.0
streamlit>=1.31.0
